added debug breakpoint

+++ train.py
@@ -179,7 +179,8 @@ def train_model(config, train_batches, validation_batches, embedding_model, numb
 					validation_model.x: validation_batches[batch], validation_model.lens: validation_lens[batch], validation_model.y: validation_labels[batch]})
 				validation_loss += loss
 				accuracy += acc
-				precision += metrics.precision_score(np.array(validation_batches[batch]), hpl, average = "micro")
+				import pdb; pdb.set_trace()
+				precision += metrics.precision_score(np.array(validation_batches[batch]).astype(np.int32), hpl, average = "micro")
 				recall += metrics.recall_score(np.array(validation_batches[batch]), hpl, average = "micro")
 				f1 += metrics.f1_score(np.array(validation_batches[batch]), hpl, average = "micro")
